+++ quirklike/Assets/AI/Standard Scripts/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable 
 {
-    void TakeDamage(int Damage);
+    void TakeDamage(float Damage);
     Transform GetTransform();
 }
